+++ lib/config.sh
@@ -12,22 +12,22 @@ typo3PublicDirectory="${composerDirectory}public/"
 selectTypo3Version() {
   echo "Select the TYPO3 version to be installed:"
   echo "  1) TYPO3 v12.4 LTS"
-  echo "  2) TYPO3 v13.4 LTS (default)"
-  echo "  3) TYPO3 v14.3 LTS"
-  read -rp 'Option [2]: ' typo3Option
+  echo "  2) TYPO3 v13.4 LTS"
+  echo "  3) TYPO3 v14.3 LTS (default)"
+  read -rp 'Option [3]: ' typo3Option
 
   case ${typo3Option} in
   1)
     typo3Version='^12.4'
     typo3MajorVersion='12'
     ;;
-  3)
-    typo3Version='^14.3'
-    typo3MajorVersion='14'
-    ;;
-  *)
+  2)
     typo3Version='^13.4'
     typo3MajorVersion='13'
+    ;;
+  *)
+    typo3Version='^14.3'
+    typo3MajorVersion='14'
     ;;
   esac